refactor: move commands into respective packages
Instead of having all CLI commands in the cmd/ directory they are moved
into the go packages they belong to. Each package that exposes one or
more commands have a command directory with one "entrypoint" that can
include subcommands, which in turn are placed in separate files in the
command directory.