feat: Allow using name of postgres cluster in postgres psql/proxy commands
The command used to only accept application name, and that is still the only option for CloudSQL instances.
For Postgres clusters, giving the name of the cluster directly might be more user friendly. We do this by looking for a cluster with the given name if no application was found.